refactor(autonomi): added docs; mod in separate file
